+++ src/Imaging/ImagePorts.tcc
@@ -81,8 +81,8 @@ template< typename ElementType, unsigned dimension >
 void
 OutputPortImageFilter< Image< ElementType, dimension > >
 ::SetImageSize( 
-		size_t 		minimums[ dimension ], 
-		size_t 		maximums[ dimension ], 
+		int32 		minimums[ dimension ], 
+		int32 		maximums[ dimension ], 
 		float32		elementExtents[ dimension ]
 	    )
 {
